refactor: migrate files from JS to TS and add types
