+++ src/components/auth/types.ts
@@ -1,0 +1,4 @@
+import { type ReactElement } from "react";
+export interface RequireAuthProps {
+  children: ReactElement;
+}
